moved ExplodeSettingTrait to Feature subdir
